% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxaturn.R
\name{taxaturn}
\alias{taxaturn}
\title{Analyze the 'turnover' of taxa.}
\description{
Analyze the 'turnover' of taxa along a defined gradient.
}
\examples{

## ------------------------------------------------
## Method `taxaturn$new`
## ------------------------------------------------

\donttest{
data(wheat_16S)
t1 <- taxaturn$new(wheat_16S, taxa_level = "Phylum", group = "Type", ordered_group = c("S", "RS", "R"))
t1 <- taxaturn$new(wheat_16S, taxa_level = "Phylum", group = "Type", ordered_group = c("S", "RS", "R"), by_group = "Plant_ID")
}

## ------------------------------------------------
## Method `taxaturn$cal_diff`
## ------------------------------------------------

\donttest{
t1$cal_diff()
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-taxaturn-new}{\code{taxaturn$new()}}
\item \href{#method-taxaturn-cal_diff}{\code{taxaturn$cal_diff()}}
\item \href{#method-taxaturn-clone}{\code{taxaturn$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-taxaturn-new"></a>}}
\if{latex}{\out{\hypertarget{method-taxaturn-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{taxaturn$new(
  dataset,
  taxa_level = "Phylum",
  group,
  ordered_group,
  by_group = NULL,
  by_group_final = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset}}{the object of \code{\link{microtable}} class.}

\item{\code{taxa_level}}{default "Phylum"; taxonomic rank name, such as "Genus".
If the provided taxonomic name is not a colname in tax_table of input dataset, 
the function will use the features in input \code{microtable$otu_table} automatically.}

\item{\code{group}}{sample group used for the selection; a colname of input \code{microtable$sample_table}.}

\item{\code{ordered_group}}{a vector representing the elements of \code{group} parameter.}

\item{\code{by_group}}{default NULL; a colname of sample_table of input dataset used to for the analysis for each element in the \code{by_group} to get the change result.
If \code{by_group} is not the smallest unit and has repetitions in the elements, mean and sd will be calculated.}

\item{\code{by_group_final}}{default NULL; NULL or other colname of sample_table of input dataset; 
   NULL represents the output is the consistent change across all the elements in \code{by_group};
a colname of sample_table of input dataset means the consistent change is obtained by those groups instead of all the elements in \code{by_group};
Note that the by_group_final can be same with by_group, in which the final change is the result of each element in \code{by_group};'
This parameter is applied only when \code{by_group} is not \code{NULL}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{res_change} and \code{res_abund}.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\donttest{
data(wheat_16S)
t1 <- taxaturn$new(wheat_16S, taxa_level = "Phylum", group = "Type", ordered_group = c("S", "RS", "R"))
t1 <- taxaturn$new(wheat_16S, taxa_level = "Phylum", group = "Type", ordered_group = c("S", "RS", "R"), by_group = "Plant_ID")
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-taxaturn-cal_diff"></a>}}
\if{latex}{\out{\hypertarget{method-taxaturn-cal_diff}{}}}
\subsection{Method \code{cal_diff()}}{
Differential test of taxonomic abudance across groups
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{taxaturn$cal_diff(
  method = c("wilcox", "t.test")[1],
  by_group = NULL,
  by_ID = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{default "wilcox"; see the following available options:
\describe{
  \item{\strong{'wilcox'}}{Wilcoxon Rank Sum and Signed Rank Tests for all paired groups }
  \item{\strong{'t.test'}}{Student's t-Test for all paired groups}
}}

\item{\code{by_group}}{default NULL; a column of sample_table used to perform the differential test 
among groups (\code{group} parameter) for each group (\code{by_group} parameter). 
So \code{by_group} has a larger scale than \code{group} parameter in the previous function.}

\item{\code{by_ID}}{default NULL; a column of sample_table used to perform paired t test or paired wilcox test for the paired data,
such as the data of plant compartments for different plant species (ID). 
So \code{by_ID} in sample_table should be the smallest unit of sample collection without any repetition in it.}

\item{\code{...}}{parameters (except measure) passed to \code{plot_alpha} function of \code{\link{trans_alpha}} class.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{res_change}, updated in the object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\donttest{
t1$cal_diff()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-taxaturn-clone"></a>}}
\if{latex}{\out{\hypertarget{method-taxaturn-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{taxaturn$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
