% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wheat_16S}
\alias{wheat_16S}
\title{The example dataset in the mecoturn package}
\format{
An R6 class object
}
\usage{
data(wheat_16S)
}
\description{
The dataset wheat_16S is structured with microtable class for the demonstration of examples.
}
\details{
\itemize{
  \item sample_table: sample information table
  \item otu_table: species-community abundance table
  \item tax_table: taxonomic table
  \item phylo_tree: phylogenetic tree
}
}
\keyword{R6}
\keyword{microtable}
\keyword{object}
